% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccLD.R
\name{bestHits}
\alias{bestHits}
\title{Returns the indexes of the top n hits}
\usage{
bestHits(x, n)
}
\arguments{
\item{x}{A vector}

\item{n}{The number of top hits that you wish to return}
}
\value{
which(rank(1/abs(x)) <= min(n,length(x))) )
}
\description{
Returns the indexes of the top n hits
}
\note{
Used internally
}

