% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cc2plink.R
\name{cc2plink}
\alias{cc2plink}
\title{reformat a case/control block to ped/map blocks}
\usage{
cc2plink(ccblock, causalOnly = FALSE, phenos = FALSE)
}
\description{
reformat a case/control block to ped/map blocks
}

