% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ccLD}
\alias{ccLD}
\title{LD summaries of simulated case/control panels}
\usage{
ccLD(ccblock, pvblock, sig.threshold = 8, minrsq = 0)
}
\arguments{
\item{ccblock}{A case/control panel retured by diseaseSims::getCCblock}

\item{pvblock}{Single-marker association results from diseaseSims::makePVblock}

\item{sig.threshold}{The -log10(p-value) beyond which a single-marker test is considered "significant"}

\item{minrsq}{Only include causative markers in the return value if their LD with the significant marker is >= minrsq}
}
\value{
A data frame summarizing LD beteween significant markers and causative markers.  Included are the population frequencies of the mutations and the effect sizes of the causative mutations.  The data frame contains: sp = position of the significant marker.  cp = position of causative marker.  sfreq = frequency of significant mutation in general population.  cfreq = frequency of causative mutation in general population. ce = effect size of causative mutation.  rsq = squared correlation coefficient between significant and causative marker genotype.
}
\description{
LD summaries of simulated case/control panels
}

