% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diseaseSims-package.r
\docType{package}
\name{diseaseSims}
\alias{diseaseSims}
\alias{diseaseSims-package}
\title{diseaseSims: a package for processing the output of simulations and case/control panels.}
\description{
diseaseSims: a package for processing the output of simulations and case/control panels.
}

