% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCCids}
\alias{getCCids}
\title{Read case/control ids from a file at a specific position}
\usage{
getCCids(filename, offset)
}
\arguments{
\item{filename}{The file name.  Should be binary, and either uncompressed or gzip compressed.}

\item{offset}{The size in bytes where the desired record begins}
}
\value{
A list of who the controls and cases were.  These are values between 1 and N, the population size.
}
\description{
Read case/control ids from a file at a specific position
}

