% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getEsizes}
\alias{getEsizes}
\title{Read effect sizes from a file at a specific position}
\usage{
getEsizes(filename, offset)
}
\arguments{
\item{filename}{The file name.  Should be binary, and either uncompressed or gzip compressed.}

\item{offset}{The size in bytes where the desired record begins}
}
\description{
Read effect sizes from a file at a specific position
}

