% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPheno}
\alias{getPheno}
\title{Read case/control panel from a file}
\usage{
getPheno(filename, offset)
}
\arguments{
\item{filename}{The file name.  Should be binary, and either uncompressed or gzip compressed.}

\item{offset}{The size in bytes where the desired record begins}
}
\description{
Read case/control panel from a file
}

