% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRiskVariantMatrix.R
\name{getRiskVariantMatrix}
\alias{getRiskVariantMatrix}
\title{Return a genotype matrix based on risk variants from a population}
\usage{
getRiskVariantMatrix(popfilename, popfileOffset, recordID = 0,
  modelName = "recessive", dominance = 0)
}
\description{
Return a genotype matrix based on risk variants from a population
}

