% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRiskVariantMatrix.R
\name{getVariantMatrix}
\alias{getVariantMatrix}
\title{Return a genotype matrix based on risk variants from a population}
\usage{
getVariantMatrix(popfilename, popfileOffset, phenofilename, phenofileOffset,
  modelName = "recessive", dominance = 0, selectedOnly = TRUE)
}
\arguments{
\item{popfilename}{The (binary, gzipped) file containing a population simulated by TFL2013_ind}

\item{popfileOffset}{The offset (in bytes) of the relevant record in popfilename}

\item{phenofilename}{The (binary, gzipped) file containing phenotypes generated by TFL2013_ind (optional)}

\item{phenofileOffset}{The offset (in bytes) of the relevant record in phenofilename (optional)}

\item{modelName}{The model of gene action.  Must be one of: recessive, additive, multiplicative, or popgen}

\item{dominance}{The dominance of a risk mutation.  Only applies to the "popgen" model of gene action}
}
\value{
genos, which is an data frame of genotypes. Rows = individuals. The first column contains trait
values for each individual, and the remaining columns = 0,1,2 copies of risk mutation

esizes, which is a numeric vector of effect sizes for each mutation in genos

positions, which are the positions of every matrix in genos
}
\description{
Return a genotype matrix based on risk variants from a population
}
\details{
The order of the columns in "genos" is in descending order of both frequency and absolute value of effect size
If phenofilename and phenofileOffset are both provided, then "traits" corresponds to individual phenotypes.
If these two arguments are NOT provided, then "traits" corresponds to the genetic component of trait value,
which is calculated by the parameter based to modelName.
Risk variant frequencies are calculated by colSums(genos)[-1]/(2*nrow(genos)), in case you want to filter
}

