% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gzLock}
\alias{gzLock}
\title{Manage access to a gz-compressed file}
\usage{
gzLock(filename, mode)
}
\arguments{
\item{filename}{The file name to write to}

\item{mode}{The mode for opening the file.}
}
\value{
An external pointer to a boostScopedLockManager
}
\description{
Manage access to a gz-compressed file
}

