% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Lock}
\alias{Lock}
\title{Manage access to a file}
\usage{
Lock(filename)
}
\arguments{
\item{filename}{The file name to write to}
}
\value{
An external pointer to a boostScopedLockManager
}
\description{
Manage access to a file
}

