% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/procIndex.R
\name{procIndex}
\alias{procIndex}
\title{Calculate offset values from index files}
\usage{
procIndex(index, column, recordno)
}
\arguments{
\item{index}{A data frame or matrix representing the index file data.}

\item{column}{The relevant column in index.}

\item{recordno}{The record number whose index we want to look up}
}
\description{
Calculate offset values from index files
}
\details{
The programs in disease_sims use file locking to allow multiple instances
of a program to write output to the same file.  In order to keep track
of particular replicates across various output files, an index file is needed.
These index files consist of at least two columns.  The first column is
an integer identifying a specific replicate.  (Typically, this number is the
same as the task ID number used in the array job to generate the data.)
The remaining columns relate to either the offset (in bytes) where the relevant
record starts (in the case of a file generated without gzip compression) or
the size of the record in bytes (for output generating using gzip compression).
This function goes through such indexes and returns the offset for the record,
allowing a seek/gzseek to be performed on an input file.
}

