% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleCCfromPop}
\alias{sampleCCfromPop}
\title{Sample a case/control panel from a population.
This is used for "on the fly" analysis in place of the make_case_control program,
for cases where the user may not want to write a case/control panel to file.}
\usage{
sampleCCfromPop(popfilename, offset, phenofilename, phenooffset, ncontrols,
  ncases, case_proportion, control_range, seed)
}
\arguments{
\item{popfilename}{The name of the population file (gzipped, binary)}

\item{offset}{The offset of the record in popfilename}

\item{phenofilename}{The name of the file recording phenotypes (gzipped, binary)}

\item{phenooffset}{The offset of the record in phenofilename}

\item{ncontrols}{The number of controls to sample}

\item{ncases}{The number of cases to sample}

\item{case_proportion}{The incidence of the disease.  Thus, an individual whose trait value is >= the 1 - case_proportion'th quantile of trait values in the population is a potential case}

\item{control_range}{A putative control individual is defined as mean +/- control_range*sd, where mean and sd refer to the distribution of the phenotype in the entire population.}

\item{seed}{Random number seed}
}
\description{
Sample a case/control panel from a population.
This is used for "on the fly" analysis in place of the make_case_control program,
for cases where the user may not want to write a case/control panel to file.
}

