% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sfs}
\alias{sfs}
\title{Site frequency spectra}
\usage{
sfs(popfile, n, seed = 0L)
}
\arguments{
\item{popfile}{The file containing the simulated populations}

\item{n}{The desired sample size.  If n == 2N, the entire population is sampled}

\item{seed}{A random number seed}
}
\value{
A data frame with 4 columns:

replicate = an arbitrary sample ID.  This is NOT the same ID number as the user may have assigned when running the simulation!!

i = the number of occurences of a mutation in the population/sample

n = the SFS for neutral variants

r = the SFS for risk variants
}
\description{
Site frequency spectra
}

