% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Unlock}
\alias{Unlock}
\title{Unlock and close a file}
\usage{
Unlock(s)
}
\arguments{
\item{s}{A boostScopedLockManager}
}
\description{
Unlock and close a file
}
\note{
Failing to call this function may cause file locks to persist until the R session ends
}

