% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vpv1.R
\name{vpv1}
\alias{vpv1}
\title{Calculate additive variation explained as a function of mutation frequency}
\usage{
vpv1(data, ofilename, replicateID, append = FALSE)
}
\arguments{
\item{data}{A list that is the return value from getRiskVariantMatrix}

\item{ofilename}{Output file name.  (Optional)}

\item{replicateID}{A unique identifier for this data set (unsigned integers, please!!)}

\item{append}{Append to output file?}
}
\value{
fitted.vals, which is a list of fitted values for each marker

mm, which is a 3-column matrix: p, variance explained by all markers with frequency <= p, based on R^2, and then variance explained by all markers with frequency <= p, based on adjusted R^2
}
\description{
Calculate additive variation explained as a function of mutation frequency
}
\details{
Output to the access file is managed via a POSIX file lock.  This lock scheme allows you to analyze multiple data sets
using your cluster, and then write all the output to a single file.  To make this work, though, be sure that append == TRUE,
otherwise you'll lock the file and then over-write it with each process!
}

