% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writeDataFrame}
\alias{writeDataFrame}
\title{Write a data frame to file}
\usage{
writeDataFrame(s, locker, colnames = TRUE, sep = "\\t")
}
\arguments{
\item{s}{A data frame}

\item{locker}{A boostScopedLockManager}

\item{colnames}{If TRUE, write column names to file}

\item{sep}{column separater}
}
\description{
Write a data frame to file
}

