% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writeMatrix}
\alias{writeMatrix}
\title{Write a matrix to file}
\usage{
writeMatrix(s, locker, colnames = TRUE, sep = "\\t")
}
\arguments{
\item{s}{A matrix}

\item{locker}{A boostScopedLockManager}

\item{colnames}{If TRUE, write column names to file}

\item{sep}{column separater}
}
\description{
Write a matrix to file
}

