% Generated by roxygen2 (4.0.1): do not edit by hand
\name{writePVblock}
\alias{writePVblock}
\title{Write the results of an association test to a file}
\usage{
writePVblock(outfilename, indexfilename, recordno, pvblock, append = TRUE,
  gzip = FALSE)
}
\arguments{
\item{outfilename}{Output file name}

\item{indexfilename}{Index file name for the output}

\item{recordno}{An unsigned integer to use as a label for pvblock.}

\item{pvblock}{A data frame. The return value of \link[diseaseSims]{makePVblock}.}

\item{append}{Boolean: append to output files or not?}

\item{gzip}{Boolean: write compressed data or not?}
}
\description{
Write the results of an association test to a file
}
\details{
If append == TRUE, then file output is handled by low-level
POSIX file locking via the C header
<fcntl.h>.  In practice, most file systems support this, but if you get
garbled output, yours may not.  The index file is locked, then the main
output is written and outfilename is closed.  The index data are then written,
and file locks are then released.

If append == FALSE, then no file locking is done, which means
different R processes should write to different output files,
otherwise you'll be overwriting your data!!!!
}

