% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ccpvals}
\alias{ccpvals}
\title{Obtain single-marker p-values for case/control data}
\usage{
ccpvals(x, status, model = "additive")
}
\arguments{
\item{x}{A matrix of genotypes, coded as number of copies of minor allele}

\item{status}{An array of 0 = control, 1 = case.  length(status) must equal nrow(x)}

\item{model}{One of "additive","recessive", or "dominant"}
}
\description{
Obtain single-marker p-values for case/control data
}

