% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{diseaseSims}
\alias{diseaseSims}
\alias{diseaseSims-package}
\title{diseaseSims: a package for processing the output of simulations and case/control panels.}
\description{
diseaseSims: a package for processing the output of simulations and case/control panels.
}

