% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makePVblock}
\alias{makePVblock}
\title{Perform single-marker association test on simulated case/control data}
\usage{
makePVblock(ccdata, esizes, status, model = "additive")
}
\arguments{
\item{ccdata}{A case/control data set returned from \link[diseaseSims]{getCCblock}}

\item{esizes}{A matrix summarizing mutation effect sizes returned from \link[diseaseSims]{getEsizes}}

\item{status}{An array specifying discrete phenotypes (0=control,1=case).  Length status must equal nrow(ccdata$geno)}

\item{model}{Model for the association test.  Must be one of additive, recessive, or domimant}
}
\value{
A data frame containing the following columns:
pos = the position of the mutation,
esize = the effect size of the mutation,
mfcontrols = minor allele frequency in controls,
mfcases = minor allele frequency in cases,
popfreq = DERIVED allele frequency in entire population (NOT minor allele frequency!!!)
score = -log10(logistic regression p-value)
}
\description{
Perform single-marker association test on simulated case/control data
}
\examples{
#This is a non-working example!
library(diseaseSims)
fn="ccfilename.bin.gz"
ccdata=getCCblock(fn,offset)
fn2="esizefile.bin.gz"
esizes=getEsizes(fn2,offset)
#Run test under additive model
pvblock=makePVblock( ccdata,esizes,c(rep(0,ccdata$ncontrols),rep(1,ccdata$ncases)) )
#Manhattan plot
plot( pvblock$pos,pvblock$score,xlab="Position",ylab=expression(paste("-log",[10],"(P-value)")))
}

