% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vpv1.R
\name{vpv1}
\alias{vpv1}
\title{Calculate additive variation explained as a function of mutation frequency}
\usage{
vpv1(data, ofilename, replicateID, append = FALSE)
}
\arguments{
\item{data}{A list that is the return value from getRiskVariantMatrix}

\item{ofilename}{Output file name.  (Optional)}
}
\value{
fitted.vals, which is a list of fitted values for each marker

mm, which is a 3-column matrix: p, variance explained by all markers with frequency <= p, based on R^2, and then variance explained by all markers with frequency <= p, based on adjusted R^2
}
\description{
Calculate additive variation explained as a function of mutation frequency
}
\details{
This is a general method that goes beyond the simple formula presented in the literature, which is based on additve models
}

