% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vpv1aov.R
\name{vpv1aov}
\alias{vpv1aov}
\title{Fit linear models to genotypes}
\usage{
vpv1aov(data, method = "lm", chunksize = 5000)
}
\arguments{
\item{data}{A list returned by getRiskVariantMatrix}

\item{method}{Method to use for model fitting.  Must be one of "biglm","slm", or "lm", corresponding to the biglm, SparseM, and base packages, respectively.}

\item{chunksize}{If the data set is massive, we'll use biglm::biglm for the regression, breaking data up into sets of chunksize rows.  In our testing, larger chunk sizes result in lower run-times.}
}
\value{
A data frame with 3 columns: allele frequency, and then estimates of the total variance in the trait explained by markers are <= that frequency.  The estimates make up the last two columns, and are based on the R^2 and adjusted R^2 of the linear model, respectively.
}
\description{
Fit linear models to genotypes
}

